Move unlocalized strings to StringsLibrary
