+++ Packages/App/Sources/StringsLibrary/Strings+Unlocalized.swift
@@ -23,7 +23,6 @@
 //  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-import CommonLibrary
 import Foundation
 
 extension Strings {
@@ -59,29 +58,9 @@ extension Strings {
 
             public static let ipV4DNS = "1.1.1.1"
 
-            public static func ipDestination(forFamily family: Address.Family) -> String {
-                switch family {
-                case .v4:
-                    return "192.168.15.0/24"
-                case .v6:
-                    return "fdbd:dcf8:d811:af73::/64"
-                }
-            }
-
-            public static func ipAddress(forFamily family: Address.Family) -> String {
-                switch family {
-                case .v4:
-                    return "192.168.15.1"
-                case .v6:
-                    return "fdbd:dcf8:d811:af73::1"
-                }
-            }
+            public static let proxyPort = "1080"
 
             public static let mtu = "1500"
-
-            public static let proxyIPv4Address = ipAddress(forFamily: .v4)
-
-            public static let proxyPort = "1080"
 
             public static let pacURL = "http://proxy.com/pac.url"
